requirements.txt
numpy
keras
tensorflow
Pillow